Support for seting ip based on X-Real-Ip/X-Forwarded-For.
If xheaders is True in config, support the X-Real-Ip/X-Forwarded-For and X-Scheme/X-Forwarded-Proto headers, which override the remote IP and URI scheme/protocol for all requests. These headers are useful when running Tornado behind a reverse proxy or load balancer.

Add xheaders to your config.yml under tornado->server to activate. For example:

tornado:
  server:
    host: 127.0.0.1
    port: 5000
    base: ''
    xheaders: True